Switching main
